Main view code style, moved code to extensions
