+++ Wikivoyage/MainViewController.swift
@@ -7,42 +7,35 @@
 //
 
 import UIKit
-import Alamofire
-import SwiftyJSON
 import PureLayout
-import SDWebImage
 
-class MainViewController: UIViewController, UISearchBarDelegate, UITableViewDataSource, UITableViewDelegate {
+class MainViewController: UIViewController {
 
-    var topView: UIView!
-    var bottomView: UIView!
+    var topView, bottomView: UIView!
     
     var locationSearchBar: UISearchBar!
-    var searchBarDimensionConstraint: NSLayoutConstraint?
-    var searchBarEdgeConstraint: NSLayoutConstraint?
-    var searchBarTop: Bool = false
+    var searchBarDimensionConstraint, searchBarEdgeConstraint: NSLayoutConstraint?
+    var searchBarTop = false
     
     var searchButton: UIButton!
-    var searchButtonWidth: NSLayoutConstraint?
-    var searchButtonHeight: NSLayoutConstraint?
-    var searchButtonEdgeConstraint: NSLayoutConstraint?
+    var searchButtonWidth, searchButtonHeight, searchButtonEdgeConstraint: NSLayoutConstraint?
     
-    var favoriteButton: UIButton!
-    var offlineButton: UIButton!
+    var favoriteButton, offlineButton: UIButton!
     
     var resultsTable: UITableView!
-    var searchResults: [SearchResult] = []
-    
-    var didSetupContraints: Bool = false
-    
+    var searchResults = [SearchResult]()
     var lastRequestid: String!
+    
+    var didSetupContraints = false
+    
+    // MARK: - View Lifecycle
     
     override func viewDidLoad() {
         super.viewDidLoad()
         setupViews()
         setupSearchBar()
         setupSearchButton()
-        setupButtons()
+        setupOtherButtons()
         setupTable()
     }
     
@@ -51,66 +44,15 @@ class MainViewController: UIViewController, UISearchBarDelegate, UITableViewData
         resetSearchBar(locationSearchBar, animated: false)
     }
     
-    override func updateViewConstraints() {
-        
-        if !didSetupContraints {
-
-            // Align top and bottom views
-            topView.autoMatchDimension(.Width, toDimension: .Width, ofView: self.view)
-            bottomView.autoMatchDimension(.Width, toDimension: .Width, ofView: self.view)
-            let views: NSArray = [topView, bottomView]
-            views.autoDistributeViewsAlongAxis(.Vertical, alignedTo: .Vertical, withFixedSpacing: 0)
-            
-            // Align search bar at top
-            locationSearchBar.autoAlignAxisToSuperviewAxis(.Vertical)
-            locationSearchBar.autoMatchDimension(.Width, toDimension: .Width, ofView: topView)
-            locationSearchBar.autoPinToTopLayoutGuideOfViewController(self, withInset: 0)
-            
-            // Align search button
-            searchButton.autoAlignAxisToSuperviewAxis(.Vertical)
-            
-            // Align results table
-            resultsTable.autoAlignAxisToSuperviewAxis(.Vertical)
-            resultsTable.autoPinEdgeToSuperviewEdge(.Left, withInset: 0)
-            resultsTable.autoPinEdgeToSuperviewEdge(.Right, withInset: 0)
-            resultsTable.autoPinToBottomLayoutGuideOfViewController(self, withInset: 0)
-            resultsTable.autoPinEdge(.Top, toEdge: .Bottom, ofView: locationSearchBar)
-            
-            // Align buttons
-            favoriteButton.autoSetDimension(.Width, toSize: 250)
-            favoriteButton.autoAlignAxisToSuperviewAxis(.Vertical)
-            offlineButton.autoSetDimension(.Width, toSize: 250)
-            offlineButton.autoAlignAxisToSuperviewAxis(.Vertical)
-            let buttons: NSArray = [favoriteButton, offlineButton]
-            buttons.autoDistributeViewsAlongAxis(.Vertical, alignedTo: .Vertical, withFixedSpacing: 30)
-            
-            didSetupContraints = true
-        }
-        
-        searchButtonWidth?.autoRemove()
-        searchButtonHeight?.autoRemove()
-        searchButtonEdgeConstraint?.autoRemove()
-        
-        if searchBarTop {
-            searchButtonWidth = searchButton.autoMatchDimension(.Width, toDimension: .Width, ofView: self.view)
-            searchButtonHeight = searchButton.autoSetDimension(.Height, toSize: 44)
-            searchButtonEdgeConstraint = searchButton.autoPinToTopLayoutGuideOfViewController(self, withInset: 0)
-        } else {
-            searchButtonWidth = searchButton.autoSetDimension(.Width, toSize: 300)
-            searchButtonHeight = searchButton.autoSetDimension(.Height, toSize: 60)
-            searchButtonEdgeConstraint = searchButton.autoPinEdgeToSuperviewEdge(.Bottom)
-        }
-        
-        super.updateViewConstraints()
-    }
+    // MARK: - Initialization
     
     func setupViews() {
         topView = UIView.newAutoLayoutView()
         bottomView = UIView.newAutoLayoutView()
-        self.view.addSubview(topView)
-        self.view.addSubview(bottomView)
+        view.addSubview(topView)
+        view.addSubview(bottomView)
     }
-
+    
     func setupSearchBar() {
         locationSearchBar = UISearchBar.newAutoLayoutView()
         locationSearchBar.delegate = self
@@ -132,7 +74,7 @@ class MainViewController: UIViewController, UISearchBarDelegate, UITableViewData
         topView.addSubview(searchButton)
     }
     
-    func setupButtons() {
+    func setupOtherButtons() {
         favoriteButton = UIButton.buttonWithType(.System) as! UIButton
         favoriteButton.setTranslatesAutoresizingMaskIntoConstraints(false)
         favoriteButton.setTitle("Favorite Locations", forState: .Normal)
@@ -160,13 +102,67 @@ class MainViewController: UIViewController, UISearchBarDelegate, UITableViewData
         resultsTable = UITableView.newAutoLayoutView()
         resultsTable.dataSource = self
         resultsTable.delegate = self
-        resultsTable.alpha = 0
         resultsTable.registerClass(SearchResultTableViewCell.self, forCellReuseIdentifier: "TableCell")
+        
+        resultsTable.alpha = 0
         resultsTable.rowHeight = 60
+        // Start separator line at 80px
         resultsTable.separatorInset = UIEdgeInsets(top: 0, left: 80, bottom: 0, right: 0)
+        // Disable separator lines for empty cells
         resultsTable.tableFooterView = UIView(frame: CGRectZero)
-        self.view.addSubview(resultsTable)
+        
+        view.addSubview(resultsTable)
     }
+    
+    // MARK: - Layout
+    
+    override func updateViewConstraints() {
+        if !didSetupContraints {
+            topView.autoMatchDimension(.Width, toDimension: .Width, ofView: view)
+            bottomView.autoMatchDimension(.Width, toDimension: .Width, ofView: view)
+            let views = NSArray(array: [topView, bottomView])
+            views.autoDistributeViewsAlongAxis(.Vertical, alignedTo: .Vertical, withFixedSpacing: 0)
+            
+            locationSearchBar.autoAlignAxisToSuperviewAxis(.Vertical)
+            locationSearchBar.autoMatchDimension(.Width, toDimension: .Width, ofView: topView)
+            locationSearchBar.autoPinToTopLayoutGuideOfViewController(self, withInset: 0)
+            
+            searchButton.autoAlignAxisToSuperviewAxis(.Vertical)
+            
+            favoriteButton.autoSetDimension(.Width, toSize: 250)
+            favoriteButton.autoAlignAxisToSuperviewAxis(.Vertical)
+            offlineButton.autoSetDimension(.Width, toSize: 250)
+            offlineButton.autoAlignAxisToSuperviewAxis(.Vertical)
+            let buttons = NSArray(array: [favoriteButton, offlineButton])
+            buttons.autoDistributeViewsAlongAxis(.Vertical, alignedTo: .Vertical, withFixedSpacing: 30)
+            
+            resultsTable.autoAlignAxisToSuperviewAxis(.Vertical)
+            resultsTable.autoPinEdgeToSuperviewEdge(.Leading)
+            resultsTable.autoPinEdgeToSuperviewEdge(.Trailing)
+            resultsTable.autoPinToBottomLayoutGuideOfViewController(self, withInset: 0)
+            resultsTable.autoPinEdge(.Top, toEdge: .Bottom, ofView: locationSearchBar)
+            
+            didSetupContraints = true
+        }
+        
+        searchButtonWidth?.autoRemove()
+        searchButtonHeight?.autoRemove()
+        searchButtonEdgeConstraint?.autoRemove()
+        
+        if searchBarTop {
+            searchButtonWidth = searchButton.autoMatchDimension(.Width, toDimension: .Width, ofView: topView)
+            searchButtonHeight = searchButton.autoSetDimension(.Height, toSize: 44)
+            searchButtonEdgeConstraint = searchButton.autoPinToTopLayoutGuideOfViewController(self, withInset: 0)
+        } else {
+            searchButtonWidth = searchButton.autoSetDimension(.Width, toSize: 300)
+            searchButtonHeight = searchButton.autoSetDimension(.Height, toSize: 60)
+            searchButtonEdgeConstraint = searchButton.autoPinEdgeToSuperviewEdge(.Bottom)
+        }
+        
+        super.updateViewConstraints()
+    }
+    
+    // MARK: - User Interaction
     
     func searchClicked(sender: UIButton!) {
         showSearchBar(locationSearchBar)
@@ -180,207 +176,14 @@ class MainViewController: UIViewController, UISearchBarDelegate, UITableViewData
         performSegueWithIdentifier("ShowOffline", sender: sender)
     }
     
-    // Search bar delegate
-    func searchBarSearchButtonClicked(searchBar: UISearchBar) {
-        dismissKeyboard()
-        if !searchBar.text.isEmpty {
-            queryTitles(searchBar.text)
-        }
-    }
-    
-    func searchBarCancelButtonClicked(searchBar: UISearchBar) {
-        resetSearchBar(searchBar, animated: true)
-    }
-    
-    func searchBar(searchBar: UISearchBar, textDidChange searchText: String) {
-        if !searchText.isEmpty {
-            let delay = 0.3
-            let time = dispatch_time(DISPATCH_TIME_NOW, Int64(delay * Double(NSEC_PER_SEC)))
-            dispatch_after(time, dispatch_get_main_queue()) {
-                if searchText == self.locationSearchBar.text {
-                    self.queryTitles(searchText)
-                }
-            }
-        }
-    }
-    
-    // Search bar helper methods
-    func resetSearchBar(searchBar: UISearchBar, animated: Bool) {
-        searchBar.text = ""
-        dismissSearchBar(searchBar, animated: animated)
-        searchResults.removeAll()
-        resultsTable.reloadData()
-    }
-    
-    func showSearchBar(searchBar: UISearchBar) {
-        searchBarTop = true
-        
-        self.view.setNeedsUpdateConstraints()
-        self.view.updateConstraintsIfNeeded()
-        
-        UIView.animateWithDuration(0.5,
-            animations: {
-                self.view.layoutIfNeeded()
-            }, completion: { finished in
-                UIView.animateWithDuration(0.2,
-                    animations: {
-                        searchBar.alpha = 1
-                        self.resultsTable.alpha = 1
-                        self.searchButton.alpha = 0
-                    }, completion: { finished in
-                        searchBar.becomeFirstResponder()
-                    })
-            })
-    }
-    
-    func dismissSearchBar(searchBar: UISearchBar, animated: Bool) {
-        searchBarTop = false
-        searchBar.resignFirstResponder()
-        
-        if animated {
-            UIView.animateWithDuration(0.2,
-                animations: {
-                    searchBar.alpha = 0
-                    self.resultsTable.alpha = 0
-                    self.searchButton.alpha = 1
-                }, completion:  { finished in
-                    self.view.setNeedsUpdateConstraints()
-                    self.view.updateConstraintsIfNeeded()
-                    UIView.animateWithDuration(0.5,
-                        animations: {
-                            self.view.layoutIfNeeded()
-                        })
-                })
-        } else {
-            self.view.setNeedsUpdateConstraints()
-            self.view.updateConstraintsIfNeeded()
-            self.view.layoutIfNeeded()
-            searchBar.alpha = 0
-            self.resultsTable.alpha = 0
-            self.searchButton.alpha = 1
-        }
-    }
-    
-    // Table view data source
-    func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCellWithIdentifier("TableCell", forIndexPath: indexPath) as! SearchResultTableViewCell
-        
-        cell.title.text = searchResults[indexPath.row].pageTitle
-
-        let placeholder = UIImage(named: "placeholder")
-        if let thumbnailURL = searchResults[indexPath.row].thumbnailURL, url = NSURL(string: thumbnailURL) {
-            cell.thumbnail.sd_setImageWithURL(url, placeholderImage: placeholder!)
-        } else {
-            cell.thumbnail.sd_setImageWithURL(nil, placeholderImage: placeholder!)
-        }
-        
-        cell.setNeedsUpdateConstraints()
-        cell.updateConstraintsIfNeeded()
-        
-        return cell
-    }
-    
-    func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        if searchResults.isEmpty {
-            tableView.separatorStyle = .None
-        } else {
-            tableView.separatorStyle = .SingleLine
-        }
-        
-        return searchResults.count
-    }
-    
-    // Table view delegate
-    func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
-        let searchResult = searchResults[indexPath.row]
-        let result = searchResults[indexPath.row]
-        performSegueWithIdentifier("ShowWeb", sender: searchResult)
-        resultsTable.deselectRowAtIndexPath(indexPath, animated: true)
-        dismissKeyboard()
-    }
-    
-    func scrollViewWillBeginDragging(scrollView: UIScrollView) {
-        dismissKeyboard()
-    }
-    
-    // Helper functions
-    func dismissKeyboard() {
-        locationSearchBar.resignFirstResponder()
-        enableCancelButton(locationSearchBar)
-    }
-    
-    func enableCancelButton(searchBar: UISearchBar) {
-        for view in searchBar.subviews {
-            for subview in view.subviews {
-                if subview.isKindOfClass(UIButton) {
-                    let button = subview as! UIButton
-                    button.enabled = true
-                }
-            }
-        }
-    }
-    
-    func queryTitles(searchTerm: String) {
-        // Update lastRequestid
-        lastRequestid = searchTerm
-        let limit = 20
-        let size = 128
-        
-        let parameters: [String: AnyObject] = [
-            "action": "query",
-            "format": "json",
-            "requestid": searchTerm,
-            "generator": "prefixsearch",
-            "gpssearch": searchTerm,
-            "gpslimit": limit,
-            "prop": "pageimages",
-            "piprop": "thumbnail",
-            "pithumbsize": size,
-            "pilimit": limit
-        ]
-        
-        Alamofire.request(.GET, "https://en.wikivoyage.org/w/api.php", parameters: parameters).responseJSON() {
-            (_, _, data, error) in
-            if(error != nil) {
-                NSLog("Error: \(error)")
-            } else {
-                let json = JSON(data!)
-                let requestid = json["requestid"].stringValue
-                // Only update results using latest request
-                if requestid == self.lastRequestid {
-                    self.searchResults.removeAll(keepCapacity: false)
-                    let results = json["query", "pages"]
-                    
-                    for (index: String, subJson: JSON) in results {
-                        let index = subJson["index"].int
-                        let pageid = subJson["pageid"].int
-                        let title = subJson["title"].string
-                        let thumbnail = subJson["thumbnail"]["source"].string
-                        
-                        let searchResult = SearchResult(index: index!, pageId: pageid!, pageTitle: title!, thumbnailURL: thumbnail)
-                        self.searchResults.append(searchResult)
-                        self.searchResults.sort { $0.index < $1.index }
-                    }
-                    
-                    self.resultsTable.reloadSections(NSIndexSet(index: 0), withRowAnimation: UITableViewRowAnimation.None)
-                }
-            }
-        }
-    }
+    // MARK: - Navigation
     
     override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
         if segue.identifier == "ShowWeb" {
-            let destination = segue.destinationViewController as! LocationWebViewController
+            let vc = segue.destinationViewController as! LocationWebViewController
             let searchResult = sender as! SearchResult
-            destination.pageId = searchResult.pageId
-            destination.pageTitle = searchResult.pageTitle
+            vc.pageId = searchResult.pageId
+            vc.pageTitle = searchResult.pageTitle
         }
     }
-    
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
 }
-
